# restaurantapi-main/openapi/swagger-latest-team12.json
{
  "swagger": "2.0",
  "info": {
    "version": "2025-04-15T08:32:20Z",
    "title": "tm3-restaurant_api-sam5"
  },
  "host": "3rh91vhdtj.execute-api.eu-west-3.amazonaws.com",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
    }
  },
  "paths": {
...
    "/locations/{id}/feedbacks": {
      "get": {
        "tags": [
          "Feedbacks"
        ],
        "summary": "returns feedbacks for a restaurant location",
        "description": "This endpoint returns feedbacks for a restaurant location",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of the restaurant location",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "query",
            "description": "Type of feedback",
            "required": true,
            "type": "string",
            "enum": [
              "Cuisine",
              "Service"
            ]
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "type": "integer",
            "default": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "type": "integer",
            "default": 20
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "totalPages": {
                  "type": "integer"
                },
                "totalElements": {
                  "type": "integer"
                },
                "size": {
                  "type": "integer"
                },
                "content": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "rate": {
                        "type": "string"
                      },
                      "comment": {
                        "type": "string"
                      },
                      "userName": {
                        "type": "string"
                      },
                      "userAvatarUrl": {
                        "type": "string"
                      },
                      "date": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      },
                      "locationId": {
                        "type": "string"
                      }
                    }
                  }
                },
                "number": {
                  "type": "integer"
                },
                "sort": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "direction": {
                        "type": "string"
                      },
                      "nullHandling": {
                        "type": "string"
                      },
                      "ascending": {
                        "type": "boolean"
                      },
                      "property": {
                        "type": "string"
                      },
                      "ignoreCase": {
                        "type": "boolean"
                      }
                    }
                  }
                },
                "first": {
                  "type": "boolean"
                },
                "last": {
                  "type": "boolean"
                },
                "numberOfElements": {
                  "type": "integer"
                },
                "pageable": {
                  "type": "object",
                  "properties": {
                    "offset": {
                      "type": "integer"
                    },
                    "sort": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "direction": {
                            "type": "string"
                          },
                          "nullHandling": {
                            "type": "string"
                          },
                          "ascending": {
                            "type": "boolean"
                          },
                          "property": {
                            "type": "string"
                          },
                          "ignoreCase": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "paged": {
                      "type": "boolean"
                    },
                    "pageSize": {
                      "type": "integer"
                    },
                    "pageNumber": {
                      "type": "integer"
                    },
                    "unpaged": {
                      "type": "boolean"
                    }
                  }
                },
                "empty": {
                  "type": "boolean"
                }
              }
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-west-3:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-3:089718700404:function:tm3-api-handler-sam5/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy"
        }
      }
    },
    "/locations/{id}/speciality-dishes": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "returns specialty dishes for a specific location",
        "description": "This endpoint returns a list of specialty dishes that are unique or signature items for a specific restaurant location",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the location",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of specialty dishes successfully retrieved",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "string"
                  },
                  "previewImageUrl": {
                    "type": "string"
                  },
                  "chefNote": {
                    "type": "string"
                  },
                  "isSignature": {
                    "type": "boolean"
                  },
                  "isSeasonalSpecial": {
                    "type": "boolean"
                  },
                  "dishType": {
                    "type": "string",
                    "enum": [
                      "APPETIZER",
                      "MAIN_COURSE",
                      "DESSERT",
                      "DRINK"
                    ]
                  },
                  "popularity": {
                    "type": "number"
                  },
                  "availableSince": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            },
            "examples": {
              "application/json": [
                {
                  "id": "dish-567",
                  "name": "Downtown Truffle Risotto",
                  "description": "Creamy arborio rice cooked with wild mushrooms and finished with black truffle shavings",
                  "price": "$28",
                  "previewImageUrl": "https://green-and-tasty.s3.eu-central-1.amazonaws.com/img/truffle-risotto.jpg",
                  "chefNote": "Our signature dish created by Chef Marco using locally foraged mushrooms and imported Italian truffles",
                  "isSignature": true,
                  "isSeasonalSpecial": false,
                  "dishType": "MAIN_COURSE",
                  "popularity": 4.9,
                  "availableSince": "2022-05-15"
                },
                {
                  "id": "dish-568",
                  "name": "Downtown Citrus Tart",
                  "description": "Tangy lemon and lime tart with a buttery shortbread crust and torched meringue topping",
                  "price": "$12",
                  "previewImageUrl": "https://green-and-tasty.s3.eu-central-1.amazonaws.com/img/citrus-tart.jpg",
                  "chefNote": "A refreshing dessert that perfectly balances sweet and tart flavors",
                  "isSignature": true,
                  "isSeasonalSpecial": false,
                  "dishType": "DESSERT",
                  "popularity": 4.7,
                  "availableSince": "2022-06-10"
                },
                {
                  "id": "dish-569",
                  "name": "Summer Berry Gazpacho",
                  "description": "Chilled soup made with fresh summer berries, cucumber, and herbs",
                  "price": "$14",
                  "previewImageUrl": "https://green-and-tasty.s3.eu-central-1.amazonaws.com/img/berry-gazpacho.jpg",
                  "chefNote": "A unique seasonal special perfect for hot summer days",
                  "isSignature": false,
                  "isSeasonalSpecial": true,
                  "dishType": "APPETIZER",
                  "popularity": 4.6,
                  "availableSince": "2024-06-01"
                }
              ]
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "Location not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "Location with ID 'location-1' not found"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-west-3:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-3:089718700404:function:tm3-api-handler-sam5/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy"
        }
      }
    },
    "/locations/{id}/tables": {
      "get": {
        "tags": [
          "Locations"
        ],
        "summary": "returns tables for a specific location",
        "description": "This endpoint returns a list of all tables available at a specific restaurant location",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the location",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "List of tables successfully retrieved",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "tableNumber": {
                    "type": "string"
                  },
                  "capacity": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "section": {
                    "type": "string",
                    "enum": [
                      "INDOOR",
                      "OUTDOOR",
                      "BALCONY",
                      "PRIVATE"
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "AVAILABLE",
                      "OCCUPIED",
                      "RESERVED",
                      "MAINTENANCE"
                    ]
                  },
                  "shape": {
                    "type": "string",
                    "enum": [
                      "ROUND",
                      "RECTANGULAR",
                      "SQUARE",
                      "OVAL"
                    ]
                  },
                  "positionX": {
                    "type": "integer",
                    "description": "X-coordinate for table positioning on floor plan"
                  },
                  "positionY": {
                    "type": "integer",
                    "description": "Y-coordinate for table positioning on floor plan"
                  },
                  "isAccessible": {
                    "type": "boolean",
                    "description": "Whether the table is wheelchair accessible"
                  },
                  "minReservationTime": {
                    "type": "integer",
                    "description": "Minimum reservation time in minutes"
                  },
                  "maxReservationTime": {
                    "type": "integer",
                    "description": "Maximum reservation time in minutes"
                  }
                }
              }
            },
            "examples": {
              "application/json": [
                {
                  "id": "table-101",
                  "tableNumber": "1",
                  "capacity": 2,
                  "section": "INDOOR",
                  "status": "AVAILABLE",
                  "shape": "ROUND",
                  "positionX": 10,
                  "positionY": 15,
                  "isAccessible": true,
                  "minReservationTime": 60,
                  "maxReservationTime": 120
                },
                {
                  "id": "table-102",
                  "tableNumber": "2",
                  "capacity": 4,
                  "section": "INDOOR",
                  "status": "RESERVED",
                  "shape": "RECTANGULAR",
                  "positionX": 25,
                  "positionY": 15,
                  "isAccessible": true,
                  "minReservationTime": 60,
                  "maxReservationTime": 180
                },
                {
                  "id": "table-103",
                  "tableNumber": "3",
                  "capacity": 6,
                  "section": "OUTDOOR",
                  "status": "AVAILABLE",
                  "shape": "RECTANGULAR",
                  "positionX": 40,
                  "positionY": 30,
                  "isAccessible": true,
                  "minReservationTime": 90,
                  "maxReservationTime": 180
                },
                {
                  "id": "table-104",
                  "tableNumber": "4",
                  "capacity": 8,
                  "section": "PRIVATE",
                  "status": "AVAILABLE",
                  "shape": "OVAL",
                  "positionX": 60,
                  "positionY": 45,
                  "isAccessible": false,
                  "minReservationTime": 120,
                  "maxReservationTime": 240
                }
              ]
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication token is missing or invalid",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "Unauthorized"
              }
            }
          },
          "403": {
            "description": "Forbidden - User does not have permission to view tables",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "Access denied: Insufficient permissions"
              }
            }
          },
          "404": {
            "description": "Location not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "Location with ID 'location-1' not found"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-west-3:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-3:089718700404:function:tm3-api-handler-sam5/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy"
        }
      }
    },
...
  },
...
}